requirements.txt
alembic==0.9.9
altair==4.1.0
appdirs==1.4.4
appnope==0.1.0
arabic-reshaper==2.1.0
argon2-cffi==20.1.0
asgiref==3.2.10
astor==0.8.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
base58==2.1.0
black==20.8b1
bleach==3.2.1
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
chromedriver==2.24.1
click==7.1.2
colorlover==0.3.0
cufflinks==0.17.3
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.2
django-bootstrap==0.2.4
django-bootstrap-form==3.4
django-bootstrap3==14.1.0
django-braces==1.14.0
django-crispy==1.8.1
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-tables2==2.3.3
docopt==0.6.2
entrypoints==0.3
Flask==1.0.2
Flask-Compress==1.8.0
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
gunicorn==20.0.4
html5lib==1.1
httplib2==0.11.3
idna==2.6
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
jedi==0.17.2
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
lazy==1.3
Mako==1.0.7
MarkupSafe==1.1.0
misaka==2.1.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
oauth2==1.9.0.post1
oauthlib==2.0.7
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
plotly==4.10.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.15.6
ptyprocess==0.6.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.6.1
Pygments==2.7.1
pyline==0.3.16
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.7.3
python-editor==1.0.3
python-xlib==0.29
pytz==2020.1
pyzmq==19.0.2
regex==2020.9.27
reportlab==3.5.55
requests==2.18.4
requests-oauthlib==0.8.0
retrying==1.3.3
selenium==3.141.0
Send2Trash==1.5.0
simpy==4.0.1
six==1.11.0
smmap==3.0.5
sns==0.1
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
sqlparse==0.4.1
statistics==1.0.3.5
streamlit==0.78.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
traitlets==5.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.22
URLObject==2.4.3
validators==0.18.2
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wincertstore==0.2
WTForms==2.1
xhtml2pdf==0.2.5
